refactor: change the structure of the application according to the task
