+++ src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { RegisterFormComponent } from './auth/pages/register-form/register-form.component';
+import { SearchResultComponent } from './youtube/components/search-result/search-result.component';
 // import { NotFoundPageComponent } from './core/pages/not-found-page/not-found-page.component';
-import { SearchResultComponent } from './search/search-result/search-result.component';
 import { DetailsComponent } from './youtube/pages/details/details.component';
 
 const routes: Routes = [
