+++ src/app/auth/auth.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { Route, RouterModule } from '@angular/router';
 import { RegisterFormComponent } from './pages/register-form/register-form.component';
 import { FormsModule } from '@angular/forms';
+import { MaterialModule } from '../shared/material.module';
 
 export const authRoutes: Route[] = [
   { path: '', component: RegisterFormComponent },
@@ -14,6 +15,7 @@ export const authRoutes: Route[] = [
     CommonModule,
     RouterModule.forChild(authRoutes),
     FormsModule,
+    MaterialModule,
   ],
 })
 export class AuthModule { }
