+++ src/app/core/components/header/header.component.scss
@@ -1,4 +1,4 @@
-@import "../../scss/mixins.scss";
+@import "../../../../scss/mixins.scss";
 
 .wrapper {
   @include wrapper(80px, #E5E5E5);
