+++ src/app/youtube/components/search-item/search-item.component.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/mixins.scss";
+@import "../../../../scss/mixins.scss";
 
 .card {
   max-width: 300px;
