+++ src/app/youtube/components/search-result/search-result.component.scss
