+++ src/app/youtube/pipes/filter-by-word.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { ISearchItem } from '../search/models/search-item.model';
+import { ISearchItem } from '../models/search-item.model';
 
 @Pipe({
   name: 'filterByWord',
